requirements.txt
telebot~=0.0.5
pyTelegramBotAPI~=4.25.0
python-dotenv~=1.1.1
pytest~=8.4.1
psycopg2-binary~=2.9.10